requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    make upgrade
#
amqp==5.2.0
    # via kombu
asgiref==3.8.1
    # via django
asn1crypto==1.5.1
    # via snowflake-connector-python
awscli==1.33.32
    # via -r requirements/reporting.in
backports-zoneinfo[tzdata]==0.2.1 ; python_version < "3.9"
    # via
    #   -c requirements/constraints.txt
    #   celery
    #   django
    #   djangorestframework
    #   kombu
bcrypt==4.2.0
    # via paramiko
billiard==4.2.0
    # via celery
boto3==1.34.150
    # via -r requirements/reporting.in
botocore==1.34.150
    # via
    #   awscli
    #   boto3
    #   s3transfer
celery==5.3.6
    # via -r requirements/reporting.in
certifi==2024.7.4
    # via
    #   py2neo
    #   requests
    #   snowflake-connector-python
cffi==1.16.0
    # via
    #   cryptography
    #   pynacl
    #   snowflake-connector-python
charset-normalizer==3.3.2
    # via
    #   requests
    #   snowflake-connector-python
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   edx-django-utils
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via awscli
cryptography==42.0.8
    # via
    #   -r requirements/reporting.in
    #   django-fernet-fields-v2
    #   paramiko
    #   pgpy
    #   pyjwt
    #   pyopenssl
    #   snowflake-connector-python
django==4.2.14
    # via
    #   -c requirements/common_constraints.txt
    #   -r requirements/base.in
    #   django-crum
    #   django-fernet-fields-v2
    #   django-filter
    #   django-model-utils
    #   django-waffle
    #   djangorestframework
    #   drf-jwt
    #   edx-django-utils
    #   edx-drf-extensions
    #   edx-rbac
django-crum==0.7.9
    # via
    #   edx-django-utils
    #   edx-rbac
django-fernet-fields-v2==0.9
    # via -r requirements/base.in
django-filter==24.2
    # via -r requirements/base.in
django-model-utils==4.5.1
    # via
    #   -r requirements/base.in
    #   edx-rbac
django-waffle==4.1.0
    # via
    #   edx-django-utils
    #   edx-drf-extensions
djangorestframework==3.15.2
    # via
    #   djangorestframework-csv
    #   drf-jwt
    #   edx-drf-extensions
djangorestframework-csv==3.0.2
    # via -r requirements/base.in
dnspython==2.6.1
    # via pymongo
docutils==0.16
    # via awscli
drf-jwt==1.19.2
    # via edx-drf-extensions
edx-django-utils==5.14.2
    # via
    #   -r requirements/base.in
    #   edx-drf-extensions
    #   edx-rest-api-client
edx-drf-extensions==10.3.0
    # via
    #   -r requirements/base.in
    #   edx-rbac
edx-opaque-keys==2.10.0
    # via
    #   -r requirements/base.in
    #   edx-drf-extensions
edx-rbac==1.9.0
    # via -r requirements/base.in
edx-rest-api-client==5.7.1
    # via -r requirements/base.in
factory-boy==3.3.0
    # via -r requirements/base.in
faker==26.0.0
    # via factory-boy
filelock==3.15.4
    # via snowflake-connector-python
idna==3.7
    # via
    #   requests
    #   snowflake-connector-python
interchange==2021.0.4
    # via py2neo
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kombu==5.3.7
    # via celery
monotonic==1.6
    # via py2neo
mysql-connector-python==9.0.0
    # via -r requirements/base.in
newrelic==9.12.0
    # via edx-django-utils
numpy==1.24.4
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.in
    #   pandas
packaging==24.1
    # via
    #   py2neo
    #   snowflake-connector-python
pandas==2.0.3
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.in
pansi==2020.7.3
    # via py2neo
paramiko==3.4.0
    # via -r requirements/reporting.in
pbr==6.0.0
    # via stevedore
pgpy==0.6.0
    # via -r requirements/reporting.in
platformdirs==4.2.2
    # via snowflake-connector-python
prompt-toolkit==3.0.47
    # via click-repl
psutil==6.0.0
    # via edx-django-utils
py2neo @ https://github.com/overhangio/py2neo/releases/download/2021.2.3/py2neo-2021.2.3.tar.gz
    # via -r requirements/reporting.in
pyasn1==0.6.0
    # via
    #   pgpy
    #   rsa
pycparser==2.22
    # via cffi
pygments==2.18.0
    # via py2neo
pyjwt[crypto]==2.8.0
    # via
    #   drf-jwt
    #   edx-drf-extensions
    #   edx-rest-api-client
    #   snowflake-connector-python
pyminizip==0.2.6
    # via -r requirements/reporting.in
pymongo==4.8.0
    # via edx-opaque-keys
pynacl==1.5.0
    # via
    #   edx-django-utils
    #   paramiko
pyopenssl==24.2.1
    # via snowflake-connector-python
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   celery
    #   faker
    #   pandas
    #   vertica-python
pytz==2024.1
    # via
    #   interchange
    #   pandas
    #   snowflake-connector-python
pyyaml==6.0.1
    # via awscli
requests==2.32.3
    # via
    #   -r requirements/base.in
    #   edx-drf-extensions
    #   edx-rest-api-client
    #   slumber
    #   snowflake-connector-python
rsa==4.7.2
    # via awscli
rules==3.4
    # via -r requirements/base.in
s3transfer==0.10.2
    # via
    #   awscli
    #   boto3
semantic-version==2.10.0
    # via edx-drf-extensions
six==1.16.0
    # via
    #   edx-rbac
    #   interchange
    #   pansi
    #   py2neo
    #   python-dateutil
    #   vertica-python
slumber==0.7.1
    # via edx-rest-api-client
snowflake-connector-python==3.12.0
    # via -r requirements/reporting.in
sortedcontainers==2.4.0
    # via snowflake-connector-python
sqlparse==0.5.1
    # via django
stevedore==5.2.0
    # via
    #   edx-django-utils
    #   edx-opaque-keys
tomlkit==0.13.0
    # via snowflake-connector-python
typing-extensions==4.12.2
    # via
    #   asgiref
    #   edx-opaque-keys
    #   kombu
    #   snowflake-connector-python
tzdata==2024.1
    # via
    #   backports-zoneinfo
    #   celery
    #   pandas
unicodecsv==0.14.1
    # via -r requirements/reporting.in
urllib3==1.26.19
    # via
    #   -c requirements/constraints.txt
    #   botocore
    #   py2neo
    #   requests
    #   snowflake-connector-python
vertica-python==1.4.0
    # via -r requirements/reporting.in
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
